CtaTracker: Fixed a typo, changed the add button to "Add route" instead of "Add track", for consistency.
Calendar: Moved the imgs/ directory up one level (so now it's components/imgs/) so that multiple apps can use it for consistency. Changed all reference from "imgs/" to "../imgs/".

Notes App: Added the "+" and "-" images to the add and remove buttons for consistency. Recommend doing this for all add and remove buttons.